requirements.txt
ipython
gunicorn
gevent
hug
pyyaml
python-etcd